# Faster encoding for lengths in BufferBackend.