--- a/src/backend/buffer.rs
+++ b/src/backend/buffer.rs
@@ -46,6 +46,32 @@ pub struct BufferBackend<S = DefaultSymbol> {
     marker: PhantomData<fn() -> S>,
 }
 
+/// We use a special encoding for the length of the string, as follows:
+///
+/// length<=254: 1-byte length, then payload
+/// length>=255: 0xFF, <usize length>, 0xFF, then payload
+///
+/// This encoding has the following properties:
+/// * strings of length<=254 use the shortest possible encoding (1 byte length)
+/// * for strings of length>=255, the encoding is sometimes longer than e.g.
+///   varlen (LEB128) encoding, but the relative overhead is small: on a 64-bit
+///   machine the worst case is a length-255 string, which would take 2+255=257
+///   bytes with, and which takes 10+255=265 bytes with this encoding. So
+///   the overhead relative to LEB128 is bounded to at most 3.1%. On 32-bit or
+///   smaller machines the overhead is even lower.
+/// * it can be branchlessly decoded in 4 instructions on x86:
+///     mov rdx, byte ptr [rdi-9]   // load the 8-byte length
+///     movzx eax, byte ptr [rdi-1] // load the 1-byte length
+///     cmp al, 0xFF
+///     cmove rax, rdx              // select between the 1-byte and 8-byte lengths
+/// * it can be decoded forwards (starting at the lowest byte) and backwards
+///   (starting at the highest byte). We decode forwards when iterating over the
+///   table. We decode backwards when resolving a symbol to a string.
+///
+/// To enable the branchless decoding of strings without under-run on the buffer,
+/// we ensure that at least 10 bytes are present in the buffer before the beginning
+/// of the first payload, by padding the beginning of the buffer as necessary.
+
 impl<S> PartialEq for BufferBackend<S>
 where
     S: Symbol,
@@ -78,34 +104,45 @@ impl<S> Default for BufferBackend<S> {
     }
 }
 
+const SHORT_LEN: usize = 1;
+const LONG_LEN: usize = mem::size_of::<usize>();
+const LEN_LOOKBEHIND: usize = LONG_LEN + SHORT_LEN;
+const MAX_SHORT_LEN: usize = 254;
+const SENTINEL_SHORT_LEN: u8 = 255;
+
+#[inline]
+fn decode_len_backwards(bytes: &[u8; LEN_LOOKBEHIND]) -> usize {
+    let short_len = bytes[LEN_LOOKBEHIND - SHORT_LEN] as usize;
+    let long_len_bytes = <&[u8; LONG_LEN]>::try_from(&bytes[..LONG_LEN]).unwrap();
+    let long_len = usize::from_le_bytes(*long_len_bytes);
+    if short_len == SENTINEL_SHORT_LEN as usize{
+        long_len
+    } else {
+        short_len
+    }
+}
+
 impl<S> BufferBackend<S>
 where
     S: Symbol,
 {
-    /// Returns the next available symbol.
-    #[inline]
-    fn next_symbol(&self) -> S {
-        expect_valid_symbol(self.buffer.len())
-    }
-
     /// Resolves the string for the given symbol if any.
     ///
-    /// # Note
-    ///
-    /// Returns the string from the given index if any as well
-    /// as the index of the next string in the buffer.
-    fn resolve_index_to_str(&self, index: usize) -> Option<(&str, usize)> {
-        let bytes = self.buffer.get(index..)?;
-        let (str_len, str_len_bytes) = decode_var_usize(bytes)?;
-        let index_str = index + str_len_bytes;
-        let str_bytes = self.buffer.get(index_str..index_str + str_len)?;
+    /// The given index is a "backwards" index.
+    fn resolve_index_to_str(&self, index: usize) -> Option<&str> {
+        let len_bytes = self.buffer.get(index - LEN_LOOKBEHIND..index)?;
+        let len_bytes = <&[u8; LEN_LOOKBEHIND]>::try_from(len_bytes).unwrap();
+        let len = decode_len_backwards(len_bytes);
+        let str_bytes = self.buffer.get(index..index + len)?;
         // SAFETY: It is guaranteed by the backend that only valid strings
         //         are stored in this portion of the buffer.
         let string = unsafe { str::from_utf8_unchecked(str_bytes) };
-        Some((string, index_str + str_len))
+        Some(string)
     }
 
     /// Resolves the string for the given symbol.
+    ///
+    /// The given index is a "backwards" index.
     ///
     /// # Note
     ///
@@ -118,26 +155,14 @@ where
     unsafe fn resolve_index_to_str_unchecked(&self, index: usize) -> &str {
         // SAFETY: The function is marked unsafe so that the caller guarantees
         //         that required invariants are checked.
-        let slice_len = unsafe { self.buffer.get_unchecked(index..) };
-        // SAFETY: The function is marked unsafe so that the caller guarantees
-        //         that required invariants are checked.
-        let (str_len, str_len_bytes) = unsafe { decode_var_usize_unchecked(slice_len) };
-        let start_str = index + str_len_bytes;
-        let str_bytes =
-            // SAFETY: The function is marked unsafe so that the caller guarantees
-            //         that required invariants are checked.
-            unsafe { self.buffer.get_unchecked(start_str..start_str + str_len) };
+        let len_bytes =
+            unsafe { self.buffer.get_unchecked(index - LEN_LOOKBEHIND..index) };
+        let len_bytes = <&[u8; LEN_LOOKBEHIND]>::try_from(len_bytes).unwrap();
+        let len = decode_len_backwards(len_bytes);
+        let str_bytes = unsafe { self.buffer.get_unchecked(index..index + len) };
         // SAFETY: It is guaranteed by the backend that only valid strings
         //         are stored in this portion of the buffer.
         unsafe { str::from_utf8_unchecked(str_bytes) }
-    }
-
-    /// Pushes the given value onto the buffer with `var7` encoding.
-    ///
-    /// Returns the amount of `var7` encoded bytes.
-    #[inline]
-    fn encode_var_usize(&mut self, value: usize) -> usize {
-        encode_var_usize(&mut self.buffer, value)
     }
 
     /// Pushes the given string into the buffer and returns its span.
@@ -146,11 +171,22 @@ where
     ///
     /// If the backend ran out of symbols.
     fn push_string(&mut self, string: &str) -> S {
-        let symbol = self.next_symbol();
-        let str_len = string.len();
-        let str_bytes = string.as_bytes();
-        self.encode_var_usize(str_len);
-        self.buffer.extend(str_bytes);
+        if self.buffer.is_empty() {
+            // Ensure at least LONG_LEN bytes are present in the buffer, so that LEN_LOOKBEHIND works.
+            self.buffer.resize(LONG_LEN, 0);
+        }
+        if string.len() <= MAX_SHORT_LEN {
+            self.buffer.reserve(SHORT_LEN + string.len());
+            self.buffer.push(string.len() as u8);
+        } else {
+            self.buffer
+                .reserve(SHORT_LEN + LONG_LEN + SHORT_LEN + string.len());
+            self.buffer.push(SENTINEL_SHORT_LEN);
+            self.buffer.extend_from_slice(&string.len().to_ne_bytes());
+            self.buffer.push(SENTINEL_SHORT_LEN);
+        }
+        let symbol = expect_valid_symbol(self.buffer.len());
+        self.buffer.extend_from_slice(string.as_bytes());
         self.len_strings += 1;
         symbol
     }
@@ -164,11 +200,9 @@ where
 
     #[cfg_attr(feature = "inline-more", inline)]
     fn with_capacity(capacity: usize) -> Self {
-        /// We encode the `usize` string length into the buffer as well.
-        const LEN_USIZE: usize = mem::size_of::<usize>();
         /// According to google the approx. word length is 5.
         const DEFAULT_STR_LEN: usize = 5;
-        let bytes_per_string = DEFAULT_STR_LEN * LEN_USIZE;
+        let bytes_per_string = DEFAULT_STR_LEN + SHORT_LEN;
         Self {
             len_strings: 0,
             buffer: Vec::with_capacity(capacity * bytes_per_string),
@@ -184,7 +218,6 @@ where
     #[inline]
     fn resolve(&self, symbol: Self::Symbol) -> Option<&str> {
         self.resolve_index_to_str(symbol.to_usize())
-            .map(|(string, _next_str_index)| string)
     }
 
     fn shrink_to_fit(&mut self) {
@@ -196,213 +229,6 @@ where
         // SAFETY: The function is marked unsafe so that the caller guarantees
         //         that required invariants are checked.
         unsafe { self.resolve_index_to_str_unchecked(symbol.to_usize()) }
-    }
-}
-
-/// Encodes the value using variable length encoding into the buffer.
-///
-/// Returns the amount of bytes used for the encoding.
-#[inline]
-fn encode_var_usize(buffer: &mut Vec<u8>, mut value: usize) -> usize {
-    if value <= 0x7F {
-        // Shortcut the common case for low value.
-        buffer.push(value as u8);
-        return 1
-    }
-    let mut len_chunks = 0;
-    loop {
-        let mut chunk = (value as u8) & 0x7F_u8;
-        value >>= 7;
-        chunk |= ((value != 0) as u8) << 7;
-        buffer.push(chunk);
-        len_chunks += 1;
-        if value == 0 {
-            break
-        }
-    }
-    len_chunks
-}
-
-/// Decodes from a variable length encoded `usize` from the buffer.
-///
-/// Returns the decoded value as first return value.
-/// Returns the number of decoded bytes as second return value.
-///
-/// # Safety
-///
-/// The caller has to make sure that the buffer contains the necessary
-/// bytes needed to properly decode a valid `usize` value.
-#[inline]
-unsafe fn decode_var_usize_unchecked(buffer: &[u8]) -> (usize, usize) {
-    let first = unsafe { *buffer.get_unchecked(0) };
-    if first <= 0x7F_u8 {
-        return (first as usize, 1)
-    }
-    let mut result: usize = 0;
-    let mut i = 0;
-    loop {
-        let byte = unsafe { *buffer.get_unchecked(i) };
-        let shifted = ((byte & 0x7F_u8) as usize) << ((i * 7) as u32);
-        result += shifted;
-        if (byte & 0x80) == 0 {
-            break
-        }
-        i += 1;
-    }
-    (result, i + 1)
-}
-
-/// Decodes from a variable length encoded `usize` from the buffer.
-///
-/// Returns the decoded value as first return value.
-/// Returns the number of decoded bytes as second return value.
-fn decode_var_usize(buffer: &[u8]) -> Option<(usize, usize)> {
-    if !buffer.is_empty() && buffer[0] <= 0x7F_u8 {
-        // Shortcut the common case for low values.
-        return Some((buffer[0] as usize, 1))
-    }
-    let mut result: usize = 0;
-    let mut i = 0;
-    loop {
-        let byte = *buffer.get(i)?;
-        let shifted = ((byte & 0x7F_u8) as usize).checked_shl((i * 7) as u32)?;
-        result = result.checked_add(shifted)?;
-        if (byte & 0x80) == 0 {
-            break
-        }
-        i += 1;
-    }
-    Some((result, i + 1))
-}
-
-#[cfg(test)]
-mod tests {
-    use super::{
-        decode_var_usize,
-        encode_var_usize,
-    };
-    #[cfg(not(feature = "std"))]
-    use alloc::vec::Vec;
-
-    #[test]
-    fn encode_var_usize_1_byte_works() {
-        let mut buffer = Vec::new();
-        for i in 0..2usize.pow(7) {
-            buffer.clear();
-            assert_eq!(encode_var_usize(&mut buffer, i), 1);
-            assert_eq!(buffer, [i as u8]);
-            assert_eq!(decode_var_usize(&buffer), Some((i, 1)));
-        }
-    }
-
-    #[test]
-    fn encode_var_usize_2_bytes_works() {
-        let mut buffer = Vec::new();
-        for i in 2usize.pow(7)..2usize.pow(14) {
-            buffer.clear();
-            assert_eq!(encode_var_usize(&mut buffer, i), 2);
-            assert_eq!(buffer, [0x80 | ((i & 0x7F) as u8), (0x7F & (i >> 7) as u8)]);
-            assert_eq!(decode_var_usize(&buffer), Some((i, 2)));
-        }
-    }
-
-    #[test]
-    #[cfg_attr(any(miri, tarpaulin), ignore)]
-    fn encode_var_usize_3_bytes_works() {
-        let mut buffer = Vec::new();
-        for i in 2usize.pow(14)..2usize.pow(21) {
-            buffer.clear();
-            assert_eq!(encode_var_usize(&mut buffer, i), 3);
-            assert_eq!(
-                buffer,
-                [
-                    0x80 | ((i & 0x7F) as u8),
-                    0x80 | (0x7F & (i >> 7) as u8),
-                    (0x7F & (i >> 14) as u8),
-                ]
-            );
-            assert_eq!(decode_var_usize(&buffer), Some((i, 3)));
-        }
-    }
-
-    /// Allows to split up the test into multiple fragments that can run in parallel.
-    #[cfg_attr(any(miri, tarpaulin), ignore)]
-    fn assert_encode_var_usize_4_bytes(range: core::ops::Range<usize>) {
-        let mut buffer = Vec::new();
-        for i in range {
-            buffer.clear();
-            assert_eq!(encode_var_usize(&mut buffer, i), 4);
-            assert_eq!(
-                buffer,
-                [
-                    0x80 | ((i & 0x7F) as u8),
-                    0x80 | (0x7F & (i >> 7) as u8),
-                    0x80 | (0x7F & (i >> 14) as u8),
-                    (0x7F & (i >> 21) as u8),
-                ]
-            );
-            assert_eq!(decode_var_usize(&buffer), Some((i, 4)));
-        }
-    }
-
-    #[test]
-    #[cfg_attr(any(miri, tarpaulin), ignore)]
-    fn encode_var_usize_4_bytes_01_works() {
-        assert_encode_var_usize_4_bytes(2usize.pow(21)..2usize.pow(24));
-    }
-
-    #[test]
-    #[cfg_attr(any(miri, tarpaulin), ignore)]
-    fn encode_var_usize_4_bytes_02_works() {
-        assert_encode_var_usize_4_bytes(2usize.pow(24)..2usize.pow(26));
-    }
-
-    #[test]
-    #[cfg_attr(any(miri, tarpaulin), ignore)]
-    fn encode_var_usize_4_bytes_03_works() {
-        assert_encode_var_usize_4_bytes(2usize.pow(26)..2usize.pow(27));
-    }
-
-    #[test]
-    #[cfg_attr(any(miri, tarpaulin), ignore)]
-    fn encode_var_usize_4_bytes_04_works() {
-        assert_encode_var_usize_4_bytes(2usize.pow(27)..2usize.pow(28));
-    }
-
-    #[test]
-    fn encode_var_u32_max_works() {
-        let mut buffer = Vec::new();
-        let i = u32::MAX as usize;
-        assert_eq!(encode_var_usize(&mut buffer, i), 5);
-        assert_eq!(buffer, [0xFF, 0xFF, 0xFF, 0xFF, 0x0F]);
-        assert_eq!(decode_var_usize(&buffer), Some((i, 5)));
-    }
-
-    #[test]
-    fn encode_var_u64_max_works() {
-        let mut buffer = Vec::new();
-        let i = u64::MAX as usize;
-        assert_eq!(encode_var_usize(&mut buffer, i), 10);
-        assert_eq!(
-            buffer,
-            [0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0x01]
-        );
-        assert_eq!(decode_var_usize(&buffer), Some((i, 10)));
-    }
-
-    #[test]
-    fn decode_var_fail() {
-        // Empty buffer.
-        assert_eq!(decode_var_usize(&[]), None);
-        // Missing buffer bytes.
-        assert_eq!(decode_var_usize(&[0x80]), None);
-        // Out of range encoded value.
-        // assert_eq!(
-        //     decode_var_usize(&[
-        //         0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0x03
-        //     ]),
-        //     None,
-        // );
     }
 }
 
@@ -431,7 +257,7 @@ impl<'a, S> Iter<'a, S> {
         Self {
             backend,
             yielded: 0,
-            current: 0,
+            current: LONG_LEN,
         }
     }
 }
@@ -450,14 +276,40 @@ where
 
     #[inline]
     fn next(&mut self) -> Option<Self::Item> {
-        self.backend.resolve_index_to_str(self.current).and_then(
-            |(string, next_string_index)| {
-                let symbol = S::try_from_usize(self.current)?;
-                self.current = next_string_index;
-                self.yielded += 1;
-                Some((symbol, string))
-            },
-        )
+        if self.backend.len_strings == self.yielded {
+            return None
+        }
+        // Parse length forwards, not backwards. To avoid the need for padding at the *end* of the buffer,
+        // when decoding forwards we use branchy code, not branchless.
+        let i = self.current;
+        // SAFETY: index is guaranteed valid because the iterator controls it.
+        let short_len = unsafe { *self.backend.buffer.get_unchecked(i) };
+        let (len, len_len) = if short_len == SENTINEL_SHORT_LEN {
+            // Long-string decoding.
+            self.current += 1;
+            // SAFETY: index is guaranteed valid because the iterator controls it.
+            let len_bytes = unsafe {
+                self.backend
+                    .buffer
+                    .get_unchecked(i + SHORT_LEN..i + SHORT_LEN + LONG_LEN)
+            };
+            let len = usize::from_ne_bytes(<[u8; LONG_LEN]>::try_from(len_bytes).unwrap())
+                as usize;
+            (len, SHORT_LEN + LONG_LEN + SHORT_LEN)
+        } else {
+            (short_len as usize, SHORT_LEN)
+        };
+        // SAFETY: index is guaranteed valid because the iterator controls it.
+        let str_bytes = unsafe {
+            self.backend
+                .buffer
+                .get_unchecked(i + len_len..i + len_len + len)
+        };
+        // SAFETY: payload guaranteed valid utf8, because that's how it was written to the buffer.
+        let str = unsafe { str::from_utf8_unchecked(str_bytes) };
+        self.current += len_len + len;
+        self.yielded += 1;
+        Some((expect_valid_symbol(i + len_len), str))
     }
 }