--- a/tests/tests.rs
+++ b/tests/tests.rs
@@ -228,12 +228,12 @@ macro_rules! gen_tests_for_backend {
         #[test]
         fn clone_works() {
             let mut interner = StringInterner::new();
-            assert_eq!(interner.get_or_intern("aa").to_usize(), 0);
+            let index = interner.get_or_intern("aa").to_usize();
 
             let mut cloned = interner.clone();
             assert_eq!(interner, cloned);
             // And the clone should have the same interned values
-            assert_eq!(cloned.get_or_intern("aa").to_usize(), 0);
+            assert_eq!(cloned.get_or_intern("aa").to_usize(), index);
         }
 
         #[test]