@Robbepop 2024-05-01
On my Macbook Pro M2 these are the speed-ups I see:
```
get_or_intern/fill-empty/new/BufferBackend
                        time:   [2.2334 ms 2.2388 ms 2.2447 ms]
                        thrpt:  [44.549 Melem/s 44.666 Melem/s 44.775 Melem/s]
                 change:
                        time:   [+0.7617% +1.2424% +1.7823%] (p = 0.00 < 0.05)
                        thrpt:  [-1.7510% -1.2271% -0.7559%]
                        Change within noise threshold.

get_or_intern/fill-empty/with_capacity/BufferBackend
                        time:   [1.2190 ms 1.2209 ms 1.2225 ms]
                        thrpt:  [81.801 Melem/s 81.905 Melem/s 82.031 Melem/s]
                 change:
                        time:   [+2.1111% +3.3640% +4.6335%] (p = 0.00 < 0.05)
                        thrpt:  [-4.4283% -3.2545% -2.0675%]
                        Performance has regressed.

get_or_intern/already-filled/BufferBackend
                        time:   [731.74 µs 733.87 µs 735.69 µs]
                        thrpt:  [135.93 Melem/s 136.26 Melem/s 136.66 Melem/s]
                 change:
                        time:   [-1.4299% -0.5519% +0.4034%] (p = 0.25 > 0.05)
                        thrpt:  [-0.4018% +0.5550% +1.4506%]
                        No change in performance detected.

get_or_intern_static/BufferBackend/get_or_intern
                        time:   [1.1811 µs 1.1842 µs 1.1879 µs]
                        thrpt:  [43.775 Melem/s 43.913 Melem/s 44.027 Melem/s]
                 change:
                        time:   [+1.5763% +1.9736% +2.3511%] (p = 0.00 < 0.05)
                        thrpt:  [-2.2971% -1.9354% -1.5519%]
                        Performance has regressed.

get_or_intern_static/BufferBackend/get_or_intern_static
                        time:   [1.1873 µs 1.1909 µs 1.1948 µs]
                        thrpt:  [43.521 Melem/s 43.665 Melem/s 43.798 Melem/s]
                 change:
                        time:   [+1.8508% +2.2119% +2.5798%] (p = 0.00 < 0.05)
                        thrpt:  [-2.5149% -2.1640% -1.8172%]
                        Performance has regressed.

resolve/already-filled/BufferBackend
                        time:   [129.64 µs 129.76 µs 129.90 µs]
                        thrpt:  [769.85 Melem/s 770.66 Melem/s 771.39 Melem/s]
                 change:
                        time:   [-18.586% -18.438% -18.296%] (p = 0.00 < 0.05)
                        thrpt:  [+22.393% +22.606% +22.828%]
                        Performance has improved.

get/already-filled/BufferBackend
                        time:   [622.42 µs 623.86 µs 625.20 µs]
                        thrpt:  [159.95 Melem/s 160.29 Melem/s 160.66 Melem/s]
                 change:
                        time:   [-6.4554% -6.0793% -5.7250%] (p = 0.00 < 0.05)
                        thrpt:  [+6.0727% +6.4728% +6.9008%]
                        Performance has improved.

iter/already-filled/BufferBackend
                        time:   [97.685 µs 99.065 µs 100.63 µs]
                        thrpt:  [993.74 Melem/s 1.0094 Gelem/s 1.0237 Gelem/s]
                 change:
                        time:   [-86.868% -86.658% -86.434%] (p = 0.00 < 0.05)
                        thrpt:  [+637.13% +649.50% +661.49%]
                        Performance has improved.
```

So no speed-ups for me for `get_or_intern` but visible speed-ups for `get`, `resolve` and in particular `iter` which is nice.