@Robbepop 2024-05-01
With https://github.com/Robbepop/string-interner/pull/64 we already have some heavy gains while keeping the old var-len encoding and I am not entirely sure if this PR is still adding a lot of improvements on top. @reinerp if you have time please feel free to rebase and re-evaluate. Otherwise I'd like to close this.