@Robbepop 2024-05-01
I now see the main benefit of this PR is not that encoding is faster but that it fixes https://github.com/Robbepop/string-interner/issues/49.